+++ HISTORY.md
@@ -1,5 +1,6 @@
 ## 1.0.7 (in progress)
 
+- Use only end of reads to detect 3' adapters in small RNA-seq pipeline.
 - Standardize dbSNP annotation to use vcfanno for all variant callers. Remove
   GATK custom annotations for non-GATK callers, which are not present in GATK4.
 - disambiguation: ensure BAM index present for non-split alignments